+++ sanity/sanity.cli.js
@@ -1,8 +1,9 @@
 import {defineCliConfig} from 'sanity/cli'
 
+// console.log(process.env)
 export default defineCliConfig({
   api: {
-    projectId: '',
+    projectId: process.env.SANITY_STUDIO_PROJECT_ID,
     dataset: 'production',
   },
 })