+++ vite/src/modules/dom.js
@@ -1,6 +1,6 @@
 import { Text } from "./animation/text";
-import { Track } from "../util/track";
-import { Alpha } from "./animation/alpha";
+// import { Track } from "../util/track";
+// import { Alpha } from "./animation/alpha";
 
 export class Dom {
   constructor() {
@@ -19,20 +19,6 @@ export class Dom {
         '[data-a="char"],[data-a="word"],[data-a="line"]'
       ),
     ].map((el) => new Text({ element: el }));
-
-    this.alpha = new Alpha({
-      element: document.querySelector("[data-a='alpha']"),
-    });
-
-    const track = document.querySelector("[data-track]");
-    if (track)
-      this.track = new Track({
-        element: track,
-        cb: {
-          // in: () => console.log("in"),
-          // out: () => console.log("out"),
-        },
-      });
 
     this.start();
   }