+++ webgl/ogl/src/gl/_quad.js
@@ -1,13 +1,14 @@
 import { Plane, Mesh } from "ogl";
 import Material from "./mat/_quad";
 
-export default class extends Mesh {
+export class Quad extends Mesh {
   constructor(gl, diff = null) {
-    super(gl);
-    this.gl = gl;
+    super(gl, {
+      geometry: new Plane(gl),
+      program: new Material(gl),
+    });
 
-    this.geometry = new Plane(this.gl);
-    this.program = new Material(this.gl);
+    this.gl = gl;
   }
 
   resize() {}